Update index.html
this is just a trivial edit because the page has gone down and I'm curious if an update will bring it back up. I updated my personal github page to use a custom domain, hardscrabble.net, and now maxjacobson.github.io/mapquest redirects there and 404s. I have a hunch that this update will rebuild the site and it won't redirect. here's hoping :)